+++ CHANGES.md
@@ -7,6 +7,7 @@ v1.1.0 (Month DD, YYYY)
 
 - Added `pdfioFileCreateTemporary` function (Issue #29)
 - Added `pdfioDictIterateKeys` function (Issue #31)
+- Added `pdfioContentPathEnd` function.
 - Fixed "install-shared" target (Issue #32)
 - Fixed `pdfioContentMatrixRotate` function.